microservice/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.4.2
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
alembic>=1.12.1
fastapi-cli>=0.0.6
typer>=0.9.0
